Interface and Circle Addition
Made a corosshair to use to move around the board and select a square to spawn objects in.

Added circles that serve as a decoy. Its still in the works. Plan is to have a triangle fight distracted triangles and survive, whereas when two non-distracted triangles fight, they both die. If you read this, Silas, any thoughts?